+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added an activatable refresh rate slider, running the sorting every Nth frame and the cutouts computation every Nth sort. Force a sort computation when a camera moves or rotates past a customizable threshold. (by [@Arthur-Aillet](https://github.com/Arthur-Aillet))
+
 - `GsplatCutout` component to edit the Gaussian splattings dynamically. A compute shader prepass is done before rendering that creates the order buffer, ignoring splats contained in cutout shapes and removing them from further calculations. (by [@Arthur-Aillet](https://github.com/Arthur-Aillet))
 
 - Multiple materials generated automatically to let the user define a custom render order. Max Render order defined in `GsplatSettings`. (by [@Arthur-Aillet](https://github.com/Arthur-Aillet))